login and signup pages are updated
