--- Login-SignUp/app/src/main/res/layout/activity_main.xml
@@ -1,108 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<LinearLayout xmlns:android="http://schemas.android.com/apk/res/android"
-    xmlns:app="http://schemas.android.com/apk/res-auto"
-    xmlns:tools="http://schemas.android.com/tools"
-    android:layout_width="match_parent"
-    android:layout_height="match_parent"
-    android:gravity="center_horizontal"
-    android:orientation="vertical"
-    android:background="@color/aqua"
-    android:padding="20dp"
-    tools:context=".MainActivity">
-
-    <TextView
-        android:layout_marginTop="70dp"
-        android:layout_width="wrap_content"
-        android:layout_height="wrap_content"
-        android:gravity="center"
-        android:text="@string/purposeful_community"
-        android:textAlignment="center"
-        android:textColor="@color/black"
-        android:textSize="20sp" />
-
-    <TextView
-        android:layout_marginTop="50dp"
-        android:layout_width="match_parent"
-        android:layout_height="wrap_content"
-        android:gravity="start"
-        android:text="@string/sign_in"
-        android:textColor="@color/black"
-        android:textSize="20sp" />
-    <TextView
-        android:layout_marginTop="10dp"
-        android:layout_width="match_parent"
-        android:layout_height="wrap_content"
-        android:gravity="start"
-        android:text="@string/hi_there_nice_to_see_you_again"
-        android:textColor="@color/gray"
-        android:textSize="15sp" />
-
-    <TextView
-        android:id="@+id/email"
-        android:layout_marginTop="20dp"
-        android:layout_width="match_parent"
-        android:layout_height="wrap_content"
-        android:gravity="start"
-        android:text="@string/email"
-        android:textColor="@color/red"
-        android:textSize="20sp" />
-
-    <EditText
-        android:id="@+id/inputmail"
-        android:layout_width="match_parent"
-        android:layout_height="wrap_content"
-        android:layout_marginTop="10dp"
-        android:hint="@string/enter_your_email"
-        android:inputType="textEmailAddress"
-
-        android:minHeight="48dp" />
-
-    <TextView
-        android:layout_marginTop="50dp"
-        android:layout_width="match_parent"
-        android:layout_height="wrap_content"
-        android:gravity="start"
-        android:text="@string/password"
-        android:textColor="@color/red"
-        android:textSize="20sp" />
-
-    <EditText
-        android:id="@+id/inputpassword"
-        android:layout_width="match_parent"
-        android:layout_height="wrap_content"
-        android:layout_marginTop="10dp"
-        android:hint="@string/enter_your_password"
-        android:inputType="textPassword"
-        android:minHeight="48dp" />
-
-
-    <Button
-        android:id="@+id/button"
-        android:layout_marginTop="30dp"
-        android:layout_width="match_parent"
-        android:layout_height="wrap_content"
-        android:text="@string/sign_in"
-        />
-
-
-    <Button
-        android:id="@+id/buttonForgot"
-        android:layout_width="wrap_content"
-        android:layout_height="wrap_content"
-        android:layout_gravity="start"
-        android:textSize="12sp"
-        android:layout_marginStart="0dp"
-        android:layout_marginTop="30dp"
-        android:text="@string/forgot_password" />
-
-    <Button
-        android:id="@+id/buttonSignUp"
-        android:layout_width="wrap_content"
-        android:layout_height="wrap_content"
-        android:layout_gravity="end"
-        android:textSize="12sp"
-        android:layout_marginStart="0dp"
-        android:layout_marginTop="-50dp"
-        android:text="@string/sign_up" />
-
-</LinearLayout>
