--- Login-SignUp/app/src/main/res/layout/signup.xml
@@ -1,115 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<LinearLayout xmlns:android="http://schemas.android.com/apk/res/android"
-    xmlns:app="http://schemas.android.com/apk/res-auto"
-    xmlns:tools="http://schemas.android.com/tools"
-    android:layout_width="match_parent"
-    android:layout_height="match_parent"
-
-    android:orientation="vertical"
-    android:background="@color/aqua"
-    android:padding="20dp"
-    tools:context=".MainActivity">
-
-    <TextView
-        android:layout_marginTop="70dp"
-        android:layout_width="wrap_content"
-        android:layout_height="wrap_content"
-        android:gravity="center"
-        android:text="@string/purposeful_community"
-        android:textAlignment="center"
-        android:textColor="@color/black"
-        android:textSize="20sp" />
-
-    <TextView
-        android:layout_marginTop="50dp"
-        android:layout_width="match_parent"
-        android:layout_height="wrap_content"
-        android:gravity="start"
-        android:text="@string/sign_up"
-        android:textColor="@color/black"
-        android:textSize="20sp" />
-
-    <TextView
-        android:layout_marginTop="20dp"
-        android:layout_width="match_parent"
-        android:layout_height="wrap_content"
-        android:gravity="start"
-        android:text="@string/username"
-        android:textColor="@color/red"
-        android:textSize="20sp" />
-
-    <EditText
-        android:id="@+id/inputusername"
-        android:layout_width="match_parent"
-        android:layout_height="wrap_content"
-        android:layout_marginTop="10dp"
-        android:hint="@string/enter_your_user_name"
-        android:inputType="textEmailAddress"
-        android:minHeight="48dp"
-        tools:ignore="TextContrastCheck" />
-
-
-    <TextView
-        android:layout_marginTop="20dp"
-        android:layout_width="match_parent"
-        android:layout_height="wrap_content"
-        android:gravity="start"
-        android:text="@string/email"
-        android:textColor="@color/red"
-        android:textSize="20sp" />
-
-    <EditText
-        android:id="@+id/inputmail2"
-        android:layout_width="match_parent"
-        android:layout_height="wrap_content"
-        android:layout_marginTop="10dp"
-        android:hint="@string/enter_your_email"
-        android:inputType="textEmailAddress"
-
-        android:minHeight="48dp" />
-
-    <TextView
-        android:layout_marginTop="10dp"
-        android:layout_width="match_parent"
-        android:layout_height="wrap_content"
-        android:gravity="start"
-        android:text="@string/password"
-        android:textColor="@color/red"
-        android:textSize="20sp" />
-
-    <EditText
-        android:id="@+id/inputpassword"
-        android:layout_width="match_parent"
-        android:layout_height="wrap_content"
-        android:layout_marginTop="10dp"
-        android:hint="@string/enter_your_password"
-        android:inputType="textPassword"
-        android:minHeight="48dp" />
-
-
-    <Button
-        android:id="@+id/button_signup"
-        android:layout_marginTop="30dp"
-        android:layout_width="match_parent"
-        android:layout_height="wrap_content"
-        android:text="@string/sign_up"
-        />
-
-    <TextView
-        android:layout_width="250dp"
-        android:layout_height="43dp"
-        android:layout_marginTop="10dp"
-        android:text="@string/have_an_account"
-        android:textSize="25sp" />
-
-    <Button
-        android:id="@+id/buttonLogin"
-        android:layout_width="wrap_content"
-        android:layout_height="wrap_content"
-        android:layout_gravity="right"
-        android:textSize="12sp"
-        android:layout_marginStart="0dp"
-        android:layout_marginTop="-50dp"
-        android:text="@string/sign_in" />
-
-</LinearLayout>
