+++ SignInSignUp/gradle/wrapper/gradle-wrapper.properties
@@ -1,0 +1,6 @@
+#Mon Nov 08 20:43:25 TRT 2021
+distributionBase=GRADLE_USER_HOME
+distributionUrl=https\://services.gradle.org/distributions/gradle-7.0.2-bin.zip
+distributionPath=wrapper/dists
+zipStorePath=wrapper/dists
+zipStoreBase=GRADLE_USER_HOME
